Moved encoding to new file
